pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI Boilerplate"
authors = ["Hide <padocon@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.8.1"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.40"}
PyJWT = "^2.4.0"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
celery = "^5.2.7"
gunicorn = "^20.1.0"
fastapi-event = "^0.1.3"
pythondi = "^1.2.4"
ujson = "^5.4.0"
aiomysql = "^0.1.1"
click = "^8.1.3"
greenlet = "1.1.3"
redis = "^4.3.4"
pip = "^23.2.1"
python-dotenv = "^1.0.0"
pytest-lazy-fixture = "^0.6.3"
factory-boy = "^3.3.0"
psycopg2-binary = "^2.9.7"
pytest-factoryboy = "^2.5.1"
pytest-cov = "^4.1.0"
ruff = "^0.0.290"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "F", # pyflakes
    "W", # pycodestyle
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "C90", # mccabe
    "A", # flake8-builtins
]
ignore = []

fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.flake8-quotes]
inline-quotes="single"